+++ .idea/misc.xml
@@ -5,6 +5,7 @@
     <option name="originalFiles">
       <list>
         <option value="$PROJECT_DIR$/pom.xml" />
+        <option value="$PROJECT_DIR$/kroxylicious-systemtests/pom.xml" />
       </list>
     </option>
   </component>
