+++ kroxylicious-filter-test-support/src/main/java/io/kroxylicious/test/condition/kafka/ApiMessageCondition.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.test.condition;
+package io.kroxylicious.test.condition.kafka;
 
 import java.util.function.Predicate;
