+++ kroxylicious-filter-test-support/src/test/java/io/kroxylicious/test/condition/ApiMessageConditionTest.java
@@ -15,6 +15,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
+import io.kroxylicious.test.condition.kafka.ApiMessageCondition;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ApiMessageConditionTest {
